+++ fml-vscode-extension/fml-validator/src/FmlTransformValidator.ts
@@ -9,6 +9,7 @@ import type {
     TransformParameter,
 } from "../../../helpers/fml_models";
 import {resolveTransformResultTypes} from "../../../helpers/fml_transform_signatures";
+import {lookupByTypeName as lookupByTypeNameSTU3} from "../../../helpers/models/generated/stu3";
 import {lookupByTypeName as lookupByTypeNameR4} from "../../../helpers/models/generated/r4";
 import {lookupByTypeName as lookupByTypeNameR4B} from "../../../helpers/models/generated/r4b";
 import {lookupByTypeName as lookupByTypeNameR5} from "../../../helpers/models/generated/r5";
@@ -24,6 +25,7 @@ interface VariableDescriptor {
 type TypeLookup = (typeName: string) => TypeModel | undefined;
 
 const lookups: Partial<Record<FhirVersion, TypeLookup>> = {
+    STU3: lookupByTypeNameSTU3,
     R4: lookupByTypeNameR4,
     R4B: lookupByTypeNameR4B,
     R5: lookupByTypeNameR5,